+++ nHIOS_PIC16F1828.X/Delay.h
@@ -17,7 +17,7 @@ extern "C" {
 #endif
 
  void delay_1ms(uint t);
- void delay_10ms(uint t);
+ void delay_10ms(ulong t);
  void init_fosc(void);
